Update color selection to try to be more intelligent
I'm not totally convinced this is the best way to do this. Rolling
through hues seems to work pretty well, but I fixed the lightness
into the middle and the saturation may not be the best approach
either. Things more or less look good for the moment however and
if nothing else it's a pretty clear improvement over just grabbing
colors at random.